Moving hue_range stuff into options.py
